+++ CHANGELOG.md
@@ -30,6 +30,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - API Keys
 - Message Templates
 
+## [4.79.18] - 2019-10-25
+## Added
+- CXV1-20403 - Add Reason Lists to navbar and expose it as Beta feature
+- CXV1-20427 - Add Transfer List page to beta features and expose it in navbar.
+
 ## [4.79.17] - 2019-10-24
 ## Added 
 - CXV1-20402 - Add reasons to navbar and expose it as Beta feature